Fix: set_power command sent more times than needed
When setting brightness or color the bulb must already be on, so we call
setPower before sending the brightness or color command, however
setPower itself should only send the command if the device is not
already on.

Due to a mismatch of types (1 !== true) the command was being sent even
when the bulb was already on.

Can fix #84.