Fix moonlight mode detection
As reported in #118 there is no value in the `support` array or
attribute in the initial advertisement that allows for dynamically
detecting if a device supports moonlight mode.

To work around this limitation we try to proactively send a `get_prop`
request for `active_mode` and will initialize the feature accordingly.

Fixes #118.